Reorganize code packaging, support mocking
Following changes were introduced:
  * existing code for '/note' endpoint was moved into subpackage;
  * variables in said code were renamed to allow new endpoints;
  * requirements, tests, and test workflow updated.

+++ tests/conftest.py
@@ -1,10 +1,10 @@
-import datetime
 import os
 
 import flask.testing
 import pytest
-import _pytest.fixtures
+import _pytest.fixtures, _pytest.monkeypatch
 import yaml
+from mongomock.mongo_client import MongoClient
 
 import notes
 from notes import Note
@@ -14,14 +14,23 @@ test_data = yaml.safe_load(open(data_file_path, 'r'))
 endpoint = '/notes'
 
 
+class PyMongoMock(MongoClient):
+    def init_app(self, app):
+        return super().__init__()
+
+
 @pytest.fixture()
-def app() -> flask.Flask:
+def app(monkeypatch: _pytest.monkeypatch) -> flask.Flask:
     """Create a Flask app object instance. Drops the MongoDB database used for testing.
 
+    :param monkeypatch: a pytest monkeypatch fixture.
     :return: the app instance.
     """
+    if os.getenv('MOCK_MONGO'):
+        monkeypatch.setattr(notes, 'mongo', PyMongoMock())
     yield notes.create_app('test.yaml')
-    notes.mongo.cx.drop_database(os.environ["MONGO_URI"].split('/')[-1])
+    if not os.getenv('MOCK_MONGO'):
+        notes.mongo.cx.drop_database(os.environ["MONGO_URI"].split('/')[-1])
 
 @pytest.fixture()
 def fill_notes(request: _pytest.fixtures.SubRequest, app: flask.Flask) -> list[dict]:
@@ -29,7 +38,7 @@ def fill_notes(request: _pytest.fixtures.SubRequest, app: flask.Flask) -> list[d
 
     :param request: pytest Request object instance.
     :param app: Flask application object instance.
-    :return: list of
+    :return: list of dicts corresponding to created Note documents.
     """
     contents = request.param['contents']
     for note in contents:
